Improved undo, allow define preview reader / writer
Added an option to define the reader and writer for the preview.  Defauled the writer to safe.  Improved the handling of undo as it didn't consider placement, so if you repeated the same action in a different part of the text, it would undo the prior action, instead of applying the markdown on the new text.